cabal: Add provider and rule for cabal settings
Rather than continually proliferate additional settings on
haskell_cabal_library, create a rule and provider that can hold cabal
specific settings.

Add `empty_lib` as the first such setting. When building an empty lib,
bazel will look for the package file, instead of for any object files.
This will fix several packages that are currently blacklisted. It will
also allow the use of packages with sublibraries and an empty main lib.

Currently there is no way to use this with a stack snapshot, short of
vendoring a package and adding a `haskell_cabal_args` to the vendored
build file. A simpler way is planned.